requirements.txt
gdown==4.7.1
pytest==7.4.1
pytest-cov==4.1.0
pylint==2.17.5
black==23.7.0
ipython==8.15.0
transformers[torch]==4.34.1
sentencepiece==0.1.99
nltk==3.8.1
datasets==2.14.4
accelerate==0.22.0
PyYAML==6.0.1
ensure==1.0.3
python-box==7.1.1
einops==0.6.1
gradio==3.44.1
absl-py==1.4.0
protobuf==3.20.3
astunparse==1.6.3
py7zr==0.20.6
evaluate==0.4.0
rouge_score==0.1.2
tensorboard==2.12.3
numpy==1.23.5
peft==0.5.0
bitsandbytes==0.41.1
trl==0.7.2
wandb==0.15.12
scipy==1.11.3
python-telegram-bot==20.6
langchain==0.0.321